requirements.txt
django
google-generativeai
python-dotenv
